implement deleting feature
