+++ kono-front/src/components/room/RoomPanel.js
@@ -6,12 +6,9 @@ import { NotiPanel } from 'components/noti';
 import { useFetch, useLanguages, useWindowDimension } from 'lib/hooks';
 import * as RoomAPI from 'api/room';
 import * as NotiAPI from 'api/noti';
-import storage from 'lib/browser/storage';
 import Text from 'res/texts/RoomPanel.text.json';
 
 const MAX_DISPLAY_NOTIS = 8;
-const STORAGE_KEY_NOTI_FILTER = '__NOTI_FILTER__';
-const NOTI_REFRESH_COOLDOWN = 24*60*60*1000;
 
 export default () => {
 
@@ -53,38 +50,14 @@ export default () => {
         notis,
         fetchNotis,
         , // isLoading,
-        isErrorNotis,
+        ,
         ,
     ] = useFetch([]);
-
-    const [notiFilter, setNotiFilter] = useState([]);
-    const [filteredNotis, setFilteredNotis] = useState([]);
 
     const refreshNotis = () => fetchNotis(NotiAPI.retrieve, [{ max_size: MAX_DISPLAY_NOTIS }]);
     useEffect(() => {
         fetchNotis(NotiAPI.retrieve, [{ max_size: MAX_DISPLAY_NOTIS }]);
     }, [fetchNotis]);
-
-    const refreshNotiFilter = () => {
-        const _notiFilter = storage.get(STORAGE_KEY_NOTI_FILTER);
-        if (_notiFilter)
-            setNotiFilter(_notiFilter);
-    };
-
-    const closeNoti = (closeSID) => {
-        const updatedNotiFilter = notiFilter.filter(({ sid }) => (sid !== closeSID));
-        updatedNotiFilter.push({ sid: closeSID, timestamp: Date.now() });
-        storage.set(STORAGE_KEY_NOTI_FILTER, updatedNotiFilter);
-        refreshNotiFilter();
-    };
-
-    useEffect(refreshNotiFilter, []);
-    useEffect(() => {
-        const filterNotis = ({ sid }) => notiFilter.every(({ sid: filterSID, timestamp }) => {
-            return (sid !== filterSID) || (timestamp + NOTI_REFRESH_COOLDOWN < Date.now());
-        });
-        setFilteredNotis(notis.filter(filterNotis));
-    }, [notis, notiFilter]);
 
     return (
         <div className={styles.RoomPanel}
@@ -105,9 +78,8 @@ export default () => {
                                 <>
                                     <RoomStatePanelMobile rooms={rooms} />
                                     <div className={styles.mobile_legend}>
-                                        <NotiPanel notis={filteredNotis} 
-                                            refresh={refreshNotis} 
-                                            closeNoti={closeNoti}/>
+                                        <NotiPanel notis={notis} 
+                                            refresh={refreshNotis} />
                                     </div>
                                 </>
                             )
@@ -120,9 +92,8 @@ export default () => {
                     <>
                         <RoomStatePanelDesktop rooms={rooms} highlight={highlight}/>
                         <div className={styles.desktop_legend}>
-                            <NotiPanel notis={filteredNotis} 
-                                refresh={refreshNotis} 
-                                closeNoti={closeNoti}/>
+                            <NotiPanel notis={notis} 
+                                refresh={refreshNotis} />
                             <RoomLegendPanel 
                                 isLoadingRooms={isLoadingRooms}
                                 lastUpdatedTime={lastUpdatedTime}
